hx711 is ok

+++ components/bsp/include/hx711.h
@@ -5,5 +5,5 @@
 
 
 void hx711_init();
-uint32_t hx711_read();
+float hx711_read();
 #endif
